+++ dfm_tools/observations.py
@@ -31,7 +31,9 @@ __all__ = ["ssh_catalog_subset",
            ]
 
 if os.name == "nt":
-    PDRIVE = "p:"
+    # windows drive letter should include trailing slash
+    # https://github.com/Deltares/dfm_tools/issues/1084
+    PDRIVE = "p:/"
 else:
     PDRIVE = "/p"