+++ docs/whats-new.md
@@ -1,5 +1,8 @@
 # What's new
 
+## UNRELEASED
+
+
 ## 0.33.0 (2025-01-20)
 
 ### Feat