+++ docs/whats-new.md
@@ -3,7 +3,7 @@
 ## UNRELEASED
 
 ### Fix
-- made p-drive paths for tide models and gesla3 work on linux also in [#1083](https://github.com/Deltares/dfm_tools/pull/1083)
+- made p-drive paths for tide models and gesla3 work on linux also in [#1083](https://github.com/Deltares/dfm_tools/pull/1083) and [#1085](https://github.com/Deltares/dfm_tools/pull/1085)
 
 
 ## 0.33.0 (2025-01-20)