+++ docs/whats-new.md
@@ -10,6 +10,7 @@
 
 ### Fix
 - made p-drive paths for tide models and gesla3 work on linux also in [#1083](https://github.com/Deltares/dfm_tools/pull/1083) and [#1085](https://github.com/Deltares/dfm_tools/pull/1085)
+- support for different face dimension names in `dfmt.enrich_rst_with_map` in [#1114](https://github.com/Deltares/dfm_tools/pull/1114)
 
 
 ## 0.33.0 (2025-01-20)