+++ docs/whats-new.md
@@ -1,12 +1,13 @@
 # What's new
 
-## UNRELEASED
+## 0.34.0 (2025-02-05)
 
 ### Feat
 - usage of outside time buffer in `dfmt.cmems_nc_to_ini()` so noon-centered or monthly timestamps are also supported in [#1087](https://github.com/Deltares/dfm_tools/pull/1087)
 - correct CMEMS daily mean data ("P1D-m") from midnight to noon by adding a 12-hour offset in `dfmt.download_CMEMS()` in [#1088](https://github.com/Deltares/dfm_tools/pull/1088)
 - updated cdsapi request to new format in [#1103](https://github.com/Deltares/dfm_tools/pull/1103)
 - added UHSLC backup for GSHHS data in [#1112](https://github.com/Deltares/dfm_tools/pull/1112)
+- auto download CMEMS phyc reanalysis from monthly mean dataset and auto convert to freq=M in [#622](https://github.com/Deltares/dfm_tools/issues/622)
 
 ### Fix
 - made p-drive paths for tide models and gesla3 work on linux also in [#1083](https://github.com/Deltares/dfm_tools/pull/1083) and [#1085](https://github.com/Deltares/dfm_tools/pull/1085)